requirements.txt
python-slugify==8.0.4
django==5.0.6
firebase-admin==6.2.0
python-dotenv==1.0.0
django-storages==1.13.3
django-cors-headers==4.3.1
google-cloud-storage==2.14.0
Pillow==10.3.0
djangorestframework==3.15.1